Web/Controllers/RouteConfig.cs
using System.Web.Mvc;
using System.Web.Routing;

namespace MvcTemplate.Controllers
{
    public class RouteConfig
    {
        public static void RegisterRoutes(RouteCollection routes)
        {
            routes.IgnoreRoute("{resource}.axd/{*pathInfo}");

            routes
                .MapRoute(
                    "DefaultMultilingual",
                    "{language}/{controller}/{action}/{id}",
                    new { controller = "Home", action = "Index", id = UrlParameter.Optional },
                    new { language = "lt-LT" },
                    new[] { "MvcTemplate.Controllers.Home" });

            routes
                .MapRoute(
                    "Default",
                    "{controller}/{action}/{id}",
                    new { language = "en-GB", controller = "Home", action = "Index", id = UrlParameter.Optional },
                    new { language = "en-GB" },
                    new[] { "MvcTemplate.Controllers.Home" });
        }
    }
}